// frontend/lib/generated-api/models/SeverityEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Analyst API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const SeverityEnum = {
  None: 'none',
  Low: 'low',
  Medium: 'medium',
  High: 'high',
} as const;
export type SeverityEnum = (typeof SeverityEnum)[keyof typeof SeverityEnum];

export function instanceOfSeverityEnum(value: any): boolean {
  for (const key in SeverityEnum) {
    if (Object.prototype.hasOwnProperty.call(SeverityEnum, key)) {
      if (SeverityEnum[key as keyof typeof SeverityEnum] === value) {
        return true;
      }
    }
  }
  return false;
}

export function SeverityEnumFromJSON(json: any): SeverityEnum {
  return SeverityEnumFromJSONTyped(json, false);
}

export function SeverityEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): SeverityEnum {
  return json as SeverityEnum;
}

export function SeverityEnumToJSON(value?: SeverityEnum | null): any {
  return value as any;
}

export function SeverityEnumToJSONTyped(value: any, ignoreDiscriminator: boolean): SeverityEnum {
  return value as SeverityEnum;
}